Move HTML styling helpers for blockquotes, bold, code blocks, expandable blockquotes, inline code, inline links, italics, specific code blocks, spoilers, strikethrough, Telegram emojis, and underlines to a separate namespace
