+++ src/StringEnricher/Helpers/Html/SpoilerHtml.cs
@@ -1,0 +1,45 @@
+﻿using StringEnricher.Nodes;
+using StringEnricher.Nodes.Html.Formatting;
+using StringEnricher.Nodes.Shared;
+
+namespace StringEnricher.Helpers.Html;
+
+/// <summary>
+/// Provides methods to apply spoiler styling in HTML format.
+/// Example: "&lt;tg-spoiler&gt;spoiler text&lt;/tg-spoiler&gt;"
+/// </summary>
+public static class SpoilerHtml
+{
+    /// <summary>
+    /// Applies spoiler style to the given style.
+    /// </summary>
+    /// <param name="style">The inner style to be wrapped with spoiler HTML tags.</param>
+    /// <typeparam name="T">The type of the inner style that implements <see cref="INode"/>.</typeparam>
+    /// <returns>A new instance of <see cref="SpoilerNode{TInner}"/> wrapping the provided inner style.</returns>
+    public static SpoilerNode<T> Apply<T>(T style) where T : INode =>
+        SpoilerNode<T>.Apply(style);
+
+    #region Overloads for Common Types
+
+    /// <summary>
+    /// Applies spoiler style to the given text.
+    /// </summary>
+    /// <param name="text">The text to be wrapped with spoiler HTML tags.</param>
+    /// <returns>A new instance of <see cref="SpoilerNode{TInner}"/> wrapping the provided text.</returns>
+    public static SpoilerNode<PlainTextNode> Apply(string text) =>
+        SpoilerNode<PlainTextNode>.Apply(text);
+
+    /// <summary>
+    /// Applies spoiler style to the given integer.
+    /// </summary>
+    /// <param name="integer">
+    /// The integer to be styled as spoiler.
+    /// </param>
+    /// <returns>
+    /// A new instance of <see cref="SpoilerNode{TInner}"/> containing the styled integer.
+    /// </returns>
+    public static SpoilerNode<IntegerNode> Apply(int integer) =>
+        SpoilerNode<IntegerNode>.Apply(integer);
+
+    #endregion
+}
